backend/requirements.txt
Flask
Flask[async]
google-cloud-firestore
googletrans==4.0.0-rc1
